requirements.txt
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-auth==2.22.0
python-dotenv==1.0.0
requests==2.31.0
groq>=0.1.1